backend/requirements.txt
setuptools>=65.0.0
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
psycopg2-binary==2.9.9
python-decouple==3.8
Pillow==10.0.0
django-filter==23.3
channels==4.0.0
channels-redis==4.1.0
firebase-admin==6.2.0
gunicorn==21.2.0
whitenoise==6.6.0
dj-database-url==2.1.0
